Support --bisect in SPEC_OPTS
Currently if RSpec is invoked with SPEC_OPTS="--bisect", an infinite number
of child processes is spawned. This is because the bisect feature works by
spawning child processes for each bisection run. In the normal cli usage case,
the --bisect argument is removed before the child processes are invoked. It
is not removed from the SPEC_OPTS variable, however, meaning each child
process believes it's starting a new bisection.

This commit allows Bisect::Runner to invoke its child processes with a
modified environment, removing --bisect if present. In modern rubies we can
use Open3.capture2e's optional environment argument - in older rubies and
JRuby we have to fall back on modifying the parent process's environment, then
restoring it after the child process has run.

Previously the Bisect::Runner was invoked during options parsing. However,
because SPEC_OPTS and the cli args are parsed by separate instances of the
parser, invoking the bisection during SPEC_OPTS parsing meant the cli args
were lost.

To address this I've made the OptionParser set a :bisect option, which is then
picked up in the main Runner to invoke the actual bisection. This has
increased the complexity of Runner.run, and there are other invocations
remaining in the OptionParser (e.g. --init and --help). These will be
refactored in a follow-up commit, so that the OptionParser sets a callable in
each case that can be called in the main Runner.